feat: auth page block
