+++ src/app/page.tsx
@@ -1,25 +1,20 @@
-"use client";
+import { headers } from "next/headers";
+import { redirect } from "next/navigation";
+import LpWrapper from "../components/lp/LpWrapper";
+import { auth } from "../lib/auth";
 
-import { useState } from "react";
-import CoreFunctions from "../components/lp/CoreFunctions";
-import Exprience from "../components/lp/Exprience";
-import Footer from "../components/lp/Footer";
-import Hero from "../components/lp/Hero";
-import Recommend from "../components/lp/Recommend";
-import TwoUseCases from "../components/lp/TwoUseCases";
+export default async function Page() {
+	const session = await auth.api.getSession({
+		headers: headers(),
+	});
 
-export default function HomePage() {
-	const [scrollY, setScrollY] = useState(0);
-	const changeScrollY = (value: number) => setScrollY(value);
+	if (!session) {
+		return <LpWrapper />;
+	}
 
-	return (
-		<div className="min-h-screen bg-white">
-			<Hero changeScrollY={changeScrollY} scrollY={scrollY} />
-			<TwoUseCases scrollY={scrollY} />
-			<CoreFunctions scrollY={scrollY} />
-			<Exprience scrollY={scrollY} />
-			<Recommend scrollY={scrollY} />
-			<Footer />
-		</div>
-	);
+	if (!session.user.osName) {
+		redirect("/enter/callback/welcome");
+	}
+
+	redirect(`/os/${session.user.osName}`);
 }
